+++ src/main/java/org/mule/module/mongo/MongoCloudConnector.java
@@ -30,16 +30,7 @@ import org.apache.commons.lang.Validate;
 import org.bson.types.BasicBSONList;
 import org.mule.api.ConnectionException;
 import org.mule.api.ConnectionExceptionCode;
-import org.mule.api.annotations.Configurable;
-import org.mule.api.annotations.Connect;
-import org.mule.api.annotations.ConnectionIdentifier;
-import org.mule.api.annotations.Connector;
-import org.mule.api.annotations.Disconnect;
-import org.mule.api.annotations.MetaDataSwitch;
-import org.mule.api.annotations.Mime;
-import org.mule.api.annotations.Processor;
-import org.mule.api.annotations.Transformer;
-import org.mule.api.annotations.ValidateConnection;
+import org.mule.api.annotations.*;
 import org.mule.api.annotations.display.Password;
 import org.mule.api.annotations.display.Placement;
 import org.mule.api.annotations.param.ConnectionKey;
@@ -69,7 +60,7 @@ import com.mongodb.util.JSON;
  * 
  * @author MuleSoft, inc.
  */
-@Connector(name = "mongo", schemaVersion = "2.0", friendlyName = "Mongo DB", minMuleVersion = "3.4", metaData = MetaDataSwitch.OFF)
+@Connector(name = "mongo", schemaVersion = "2.0", friendlyName = "Mongo DB", minMuleVersion = "3.5", metaData = MetaDataSwitch.OFF)
 public class MongoCloudConnector
 {
     private static final Logger LOGGER = LoggerFactory.getLogger(MongoCloudConnector.class);
@@ -78,8 +69,6 @@ public class MongoCloudConnector
     private static final String WRITE_CONCERN_DEFAULT_VALUE = "DATABASE_DEFAULT";
     private static final String BACKUP_THREADS = "5";
     private static final String DEFAULT_OUTPUT_DIRECTORY = "dump";
-
-    private final Mongo.Holder mongoClientHolder = Mongo.Holder.singleton();
 
     /**
      * The host of the Mongo server, it can also be a list of comma separated hosts for replicas
@@ -157,6 +146,7 @@ public class MongoCloudConnector
      * @return Result of the operation
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public WriteResult addUser(final String newUsername, final String newPassword)
     {
         return client.addUser(newUsername, newPassword);
@@ -168,6 +158,7 @@ public class MongoCloudConnector
      * {@sample.xml ../../../doc/mongo-connector.xml.sample mongo:drop-database}
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void dropDatabase()
     {
         client.dropDatabase();
@@ -181,6 +172,7 @@ public class MongoCloudConnector
      * @return the list of names of collections available at this database
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Collection<String> listCollections()
     {
         return client.listCollections();
@@ -195,6 +187,7 @@ public class MongoCloudConnector
      * @return if the collection exists
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public boolean existsCollection(final String collection)
     {
         return client.existsCollection(collection);
@@ -209,6 +202,7 @@ public class MongoCloudConnector
      * @param collection the name of the collection to drop
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void dropCollection(final String collection)
     {
         client.dropCollection(collection);
@@ -225,6 +219,7 @@ public class MongoCloudConnector
      * @param size the maximum size of the new collection
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void createCollection(final String collection,
                                  @Optional @Default(CAPPED_DEFAULT_VALUE) final boolean capped,
                                  @Optional final Integer maxObjects,
@@ -244,6 +239,7 @@ public class MongoCloudConnector
      * @return the id that was just insterted
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public String insertObject(final String collection,
                                @Optional @Default("#[payload]") final DBObject dbObject,
                                @Optional @Default(WRITE_CONCERN_DEFAULT_VALUE) final WriteConcern writeConcern)
@@ -266,6 +262,7 @@ public class MongoCloudConnector
      * @return the id that was just insterted
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public String insertObjectFromMap(final String collection,
                                       @Placement(group = "Element Attributes") final Map<String, Object> elementAttributes,
                                       @Optional @Default(WRITE_CONCERN_DEFAULT_VALUE) final WriteConcern writeConcern)
@@ -290,6 +287,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to update
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void updateObjects(final String collection,
                               final DBObject query,
                               @Optional @Default("#[payload]") final DBObject element,
@@ -317,6 +315,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to update
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void updateObjectsUsingQueryMap(final String collection,
                                            final Map<String, Object> queryAttributes,
                                            final DBObject element,
@@ -344,6 +343,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to update
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void updateObjectsUsingMap(final String collection,
                                       @Placement(group = "Query Attributes") final Map<String, Object> queryAttributes,
                                       @Placement(group = "Element Attributes") final Map<String, Object> elementAttributes,
@@ -371,6 +371,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to update
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void updateObjectsByFunction(final String collection,
                                         final String function,
                                         final DBObject query,
@@ -401,6 +402,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to update
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void updateObjectsByFunctionUsingMap(final String collection,
                                                 final String function,
                                                 final Map<String, Object> queryAttributes,
@@ -425,6 +427,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to persist the object
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void saveObject(final String collection,
                            @Optional @Default("#[payload]") final DBObject element,
                            @Optional @Default(WRITE_CONCERN_DEFAULT_VALUE) final WriteConcern writeConcern)
@@ -442,6 +445,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to persist the object
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void saveObjectFromMap(final String collection,
                                   @Placement(group = "Element Attributes") final Map<String, Object> elementAttributes,
                                   @Optional @Default(WRITE_CONCERN_DEFAULT_VALUE) final WriteConcern writeConcern)
@@ -462,6 +466,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to remove the object
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void removeObjects(final String collection,
                               @Optional @Default("#[payload]") final DBObject query,
                               @Optional @Default(WRITE_CONCERN_DEFAULT_VALUE) final WriteConcern writeConcern)
@@ -481,6 +486,7 @@ public class MongoCloudConnector
      * @param writeConcern the write concern used to remove the object
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void removeObjectsUsingQueryMap(final String collection,
                                     @Placement(group = "Query Attributes") @Optional final Map<String, Object> queryAttributes,
                                     @Optional @Default(WRITE_CONCERN_DEFAULT_VALUE) final WriteConcern writeConcern)
@@ -511,6 +517,7 @@ public class MongoCloudConnector
      * @return an iterable that retrieves the resulting collection of {@link DBObject}
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Iterable<DBObject> mapReduceObjects(final String collection,
                                                final String mapFunction,
                                                final String reduceFunction,
@@ -531,6 +538,7 @@ public class MongoCloudConnector
      * @return the amount of objects that matches the query
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public long countObjects(final String collection, @Optional @Default("#[payload]") final DBObject query)
     {
         return client.countObjects(collection, query);
@@ -548,6 +556,7 @@ public class MongoCloudConnector
      * @return the amount of objects that matches the query
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public long countObjectsUsingQueryMap(final String collection,
                                           @Placement(group = "Query Attributes") @Optional final Map<String, Object> queryAttributes)
     {
@@ -570,6 +579,7 @@ public class MongoCloudConnector
      * @return an iterable of {@link DBObject}
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Iterable<DBObject> findObjects(final String collection,
                                           @Optional @Default("") final DBObject query,
                                           @Placement(group = "Fields") @Optional final List<String> fields,
@@ -595,6 +605,7 @@ public class MongoCloudConnector
      * @return an iterable of {@link DBObject}
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Iterable<DBObject> findObjectsUsingQueryMap(final String collection,
                                                        @Placement(group = "Query Attributes") @Optional final Map<String, Object> queryAttributes,
                                                        @Placement(group = "Fields") @Optional final List<String> fields,
@@ -618,6 +629,7 @@ public class MongoCloudConnector
      * @return a {@link DBObject} that matches the query. If nothing matches and the failOnNotFound is set to false, null will be returned
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public DBObject findOneObject(final String collection,
                                   @Optional @Default("#[payload]") final DBObject query,
                                   @Placement(group = "Fields") @Optional final List<String> fields,
@@ -640,6 +652,7 @@ public class MongoCloudConnector
      * @return a {@link DBObject} that matches the query. If nothing matches and the failOnNotFound is set to false, null will be returned
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public DBObject findOneObjectUsingQueryMap(final String collection,
                                                @Placement(group = "Query Attributes") final Map<String, Object> queryAttributes,
                                                @Placement(group = "Fields") @Optional final List<String> fields,
@@ -659,6 +672,7 @@ public class MongoCloudConnector
      * @param order the indexing order
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void createIndex(final String collection,
                             final String field,
                             @Optional @Default("ASC") final IndexOrder order)
@@ -675,6 +689,7 @@ public class MongoCloudConnector
      * @param index the name of the index to drop
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void dropIndex(final String collection, final String index)
     {
         client.dropIndex(collection, index);
@@ -689,6 +704,7 @@ public class MongoCloudConnector
      * @return a collection of {@link DBObject} with indices information
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Collection<DBObject> listIndices(final String collection)
     {
         return client.listIndices(collection);
@@ -709,6 +725,7 @@ public class MongoCloudConnector
      * @throws IOException IOException
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public DBObject createFileFromPayload(@Payload final Object payload,
                                           final String filename,
                                           @Optional final String contentType,
@@ -751,6 +768,7 @@ public class MongoCloudConnector
      * @return a {@link DBObject} files iterable
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Iterable<DBObject> findFiles(@Optional @Default("#[payload]") final DBObject query)
     {
         return client.findFiles(from(query));
@@ -765,6 +783,7 @@ public class MongoCloudConnector
      * @return a {@link DBObject} files iterable
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Iterable<DBObject> findFilesUsingQueryMap(@Placement(group = "Query Attributes") @Optional final Map<String, Object> queryAttributes)
     {
         return client.findFiles((DBObject) adapt(queryAttributes));
@@ -780,6 +799,7 @@ public class MongoCloudConnector
      * @return a {@link DBObject}
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public DBObject findOneFile(final DBObject query)
     {
         return client.findOneFile(from(query));
@@ -795,6 +815,7 @@ public class MongoCloudConnector
      * @return a {@link DBObject}
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public DBObject findOneFileUsingQueryMap(@Placement(group = "Query Attributes") final Map<String, Object> queryAttributes)
     {
         return client.findOneFile((DBObject) adapt(queryAttributes));
@@ -810,6 +831,7 @@ public class MongoCloudConnector
      * @return an InputStream to the file contents
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public InputStream getFileContent(@Optional @Default("#[payload]") final DBObject query)
     {
         return client.getFileContent(from(query));
@@ -825,6 +847,7 @@ public class MongoCloudConnector
      * @return an InputStream to the file contents
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public InputStream getFileContentUsingQueryMap(@Placement(group = "Query Attributes") final Map<String, Object> queryAttributes)
     {
         return client.getFileContent((DBObject) adapt(queryAttributes));
@@ -840,6 +863,7 @@ public class MongoCloudConnector
      * @return an iterable of {@link DBObject}
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Iterable<DBObject> listFiles(@Optional @Default("#[payload]") final DBObject query)
     {
         return client.listFiles(from(query));
@@ -855,6 +879,7 @@ public class MongoCloudConnector
      * @return an iterable of {@link DBObject}
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public Iterable<DBObject> listFilesUsingQueryMap(@Placement(group = "Query Attributes") @Optional final Map<String, Object> queryAttributes)
     {
         return client.listFiles((DBObject) adapt(queryAttributes));
@@ -869,6 +894,7 @@ public class MongoCloudConnector
      * @param query the {@link DBObject} optional query
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void removeFiles(@Optional @Default("#[payload]") final DBObject query)
     {
         client.removeFiles(from(query));
@@ -883,6 +909,7 @@ public class MongoCloudConnector
      * @param queryAttributes the optional query
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void removeFilesUsingQueryMap(@Placement(group = "Query Attributes") @Optional final Map<String, Object> queryAttributes)
     {
         client.removeFiles((DBObject) adapt(queryAttributes));
@@ -899,6 +926,7 @@ public class MongoCloudConnector
      * @return The result of the command
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public DBObject executeCommand(final String commandName, @Optional final String commandValue)
     {
         final DBObject dbObject = fromCommand(commandName, commandValue);
@@ -922,6 +950,7 @@ public class MongoCloudConnector
      * @throws IOException if an error occurs during the dump
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void dump(@Optional @Default(DEFAULT_OUTPUT_DIRECTORY) final String outputDirectory,
                      @Optional final String outputName,
                      @Optional @Default("false") final boolean zip,
@@ -952,6 +981,7 @@ public class MongoCloudConnector
      * @throws IOException if an error occurs during the incremental dump
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void incrementalDump(@Optional @Default(DEFAULT_OUTPUT_DIRECTORY) final String outputDirectory,
                                 @Optional final String incrementalTimestampFile) throws IOException
     {
@@ -976,6 +1006,7 @@ public class MongoCloudConnector
      * @throws IOException if an error occurs during restore of the database
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void restore(@Optional @Default(DEFAULT_OUTPUT_DIRECTORY) final String inputPath,
                         @Optional @Default("false") final boolean drop,
                         @Optional @Default("false") final boolean oplogReplay) throws IOException
@@ -993,6 +1024,7 @@ public class MongoCloudConnector
      * @see <a href="http://docs.mongodb.org/ecosystem/drivers/java-concurrency/">MongoDB: Java Driver Concurrency</a>
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void startConsistentRequest() {
         client.requestStart();
     }
@@ -1005,6 +1037,7 @@ public class MongoCloudConnector
      * @see <a href="http://docs.mongodb.org/ecosystem/drivers/java-concurrency/">MongoDB: Java Driver Concurrency</a>
      */
     @Processor
+	@InvalidateConnectionOn(exception = IllegalStateException.class)
     public void endConsistentRequest() {
         client.requestDone();
     }
@@ -1088,22 +1121,21 @@ public class MongoCloudConnector
      * 
      * @param username the username to use for authentication. NOTE: Please use a dummy user if you
      *            have disabled Mongo authentication
-     * @param password the password to use for authentication. NOTE: Please use a dummy password if
-     *            you have disabled Mongo authentication
+     * @param password the password to use for authentication. If the password is null or whitespaces only the connector
+     *                 won't use authentication.
      * @param database Name of the database
      * @return the newly created {@link MongoSession}
      * @throws org.mule.api.ConnectionException
      */
     @Connect
     public void connect(@ConnectionKey final String username,
-                        @Password final String password,
-                        @ConnectionKey @Optional @Default("test") final String database) throws ConnectionException
+                        @Optional @Password final String password,
+                        @ConnectionKey final String database) throws ConnectionException
     {
         try
         {
-            // TODO: use MongoClient instead of Mongo once the Holder supports it (in driver version 2.12.0)
-            mongo = mongoClientHolder.connect(new MongoURI(getMongoClientURI(username, password, database)));
-
+            mongo = new com.mongodb.MongoClient(getMongoClientURI(username, password, database));
+            
             this.client = new MongoClientImpl(getDatabase(mongo, username, password, database));
         }
         catch (final MongoException me)
@@ -1185,7 +1217,7 @@ public class MongoCloudConnector
                 client = null;
             }
         }
-
+        
         if (mongo != null)
         {
             try
@@ -1206,7 +1238,7 @@ public class MongoCloudConnector
     @ValidateConnection
     public boolean isConnected()
     {
-        return this.client != null;
+        return this.client != null && this.mongo != null && mongo.getConnector().isOpen();
     }
 
     @ConnectionIdentifier
@@ -1221,7 +1253,7 @@ public class MongoCloudConnector
                            final String database) throws ConnectionException
     {
         final DB db = mongo.getDB(database);
-        if (password != null)
+        if (StringUtils.isNotBlank(password))
         {
             Validate.notNull(username, "Username must not be null if password is set");
             if (!db.isAuthenticated())